async action to mock-fetch projects
