+++ webv2/src/components/locale/factory.tsx
@@ -1,32 +1,33 @@
 import type { FC } from 'react';
 import { BaseDictionary, ExtractDictionaryLanguageCodes } from './types';
 import { PartialWithOneRequiredKey } from 'src/utils/typescript';
+import { useSelector } from 'react-redux';
 
-export function factory<D extends BaseDictionary>(
+export function factory<D extends BaseDictionary, S>(
   dictionary: D,
-  getLanguageCode: () => ExtractDictionaryLanguageCodes<D>,
+  getLanguageCode: (state: S) => ExtractDictionaryLanguageCodes<D>,
   fallbackText = 'MISSING_TRANSLATION',
-): FC<
-  // pass dictionary key as prop itself: eg <Locale navbar-section-contribute />
-  | PartialWithOneRequiredKey<Record<keyof D, boolean>>
-  // or pass dictionary key as string prop "localeKey": eg <Locale localeKey="navbar-section-contribute" />
-  | { localeKey: keyof D }
-> {
-  return ({ localeKey, ...localeKeys }) => {
-    const languageCode = getLanguageCode() as string;
-    const key = (localeKey || Object.keys(localeKeys)[0]) as string;
-    return localize(dictionary, languageCode, key, fallbackText);
-  };
-}
-
-export function functionFactory<D extends BaseDictionary>(
-  dictionary: D,
-  getLanguageCode: () => ExtractDictionaryLanguageCodes<D>,
-  fallbackText = 'MISSING_TRANSLATION',
-) {
-  return (key: keyof D) => {
-    const languageCode = getLanguageCode() as string;
-    return localize(dictionary, languageCode, key as string, fallbackText);
+): {
+  Locale: FC<
+    // pass dictionary key as prop itself: eg <Locale navbar-section-contribute />
+    | PartialWithOneRequiredKey<Record<keyof D, boolean>>
+    // or pass dictionary key as string prop "localeKey": eg <Locale localeKey="navbar-section-contribute" />
+    | { localeKey: keyof D }
+  >;
+  useLocale: () => { localize: (key: keyof D) => string };
+} {
+  return {
+    Locale: ({ localeKey, ...localeKeys }) => {
+      const languageCode = useSelector(getLanguageCode) as string;
+      const key = (localeKey || Object.keys(localeKeys)[0]) as string;
+      return localize(dictionary, languageCode, key, fallbackText);
+    },
+    useLocale: () => {
+      const languageCode = useSelector(getLanguageCode) as string;
+      return {
+        localize: (key: keyof D) => localize(dictionary, languageCode, key as string, fallbackText),
+      };
+    },
   };
 }
